Goresword rebalanced
Heavy combo multipliers reduced from 2.3>2.4>2.7 to 2.1>2.2>2.5
Crit Chance brought down from 10% to 5%

Brought down heavy combo multipliers to be more in line with hammer multipliers and make the horizontal attacks more attractive - closed the gap in DPS between Goresword and competitors like Scythe.

Goresword already has the in-built Mantle to make up for the lack of fast attacks - it didn't need to be so far ahead of its competitors in the damage department. DPS weapons like the Scythe should now be more attractive for non-facetank builds.